+++ R/model_info.R
@@ -1502,7 +1502,7 @@ model_info.bfsl <- function(x, verbose = TRUE, ...) {
 
 #' @export
 model_info.marginaleffects <- function(x, ...) {
-  model_info(attributes(x)$model)
+  model_info(marginaleffects::components(x, "model"))
 }