+++ R/n_obs.R
@@ -49,7 +49,7 @@ n_obs.default <- function(x, ...) {
     error = function(e) {
       FALSE
     }
-  )
+    })
 
   if (isTRUE(is_binomial)) {
     return(n_obs.glm(x, ...))
@@ -69,8 +69,7 @@ n_obs.glm <- function(x, disaggregate = FALSE, ...) {
     },
     error = function(e) {
       FALSE
-    }
-  )
+    })
 
   .nobs <- stats::nobs(x)
 
@@ -468,7 +467,8 @@ n_obs.riskRegression <- function(x, ...) {
 
 #' @export
 n_obs.marginaleffects <- function(x, ...) {
-  n_obs(attributes(x)$model)
+  m <- marginaleffects::components(x, "model")
+  n_obs(m)
 }