+++ tests/testthat/test-marginaleffects.R
@@ -31,12 +31,9 @@ test_that("marginaleffects", {
   skip_if_not_installed("parameters", minimum_version = "0.24.0")
   skip_if_not_installed("marginaleffects", minimum_version = "0.24.0.6")
 
-  expect_named(
-    parameters::model_parameters(x),
-    c(
-      "rowid", "Parameter", "Comparison", "Coefficient", "SE", "Statistic", "p",
-      "S", "CI", "CI_low", "CI_high", "Predicted", "Species",
-      "Petal.Length", "Sepal.Width"
-    )
-  )
+
+  cols <- c("rowid", "Parameter", "Comparison", "Coefficient", "SE", "Statistic", "p", "S", "CI", "CI_low", "CI_high", "Predicted", "Species", "Petal.Length")
+  expect_true(all(cols %in% colnames(parameters::model_parameters(x))))
+
+  expect_equal(n_obs(x), 150) # nrow(iris)
 })